+++ src/core/operations/ParseColourCode.mjs
@@ -113,7 +113,7 @@ CMYK: ${cmyk}
     }).on('colorpickerChange', function(e) {
         var color = e.color.string('rgba');
         document.getElementById('input-text').value = color;
-        window.app.autoBake();
+        window.app.manager.input.debounceInputChange(new Event("keyup"));
     });
 </script>`;
     }