+++ src/core/operations/XORBruteForce.mjs
@@ -25,7 +25,7 @@ class XORBruteForce extends Operation {
         this.module = "Default";
         this.description = "Enumerate all possible XOR solutions. Current maximum key length is 2 due to browser performance.<br><br>Optionally enter a string that you expect to find in the plaintext to filter results (crib).";
         this.infoURL = "https://wikipedia.org/wiki/Exclusive_or";
-        this.inputType = "byteArray";
+        this.inputType = "ArrayBuffer";
         this.outputType = "string";
         this.args = [
             {
@@ -72,11 +72,12 @@ class XORBruteForce extends Operation {
     }
 
     /**
-     * @param {byteArray} input
+     * @param {ArrayBuffer} input
      * @param {Object[]} args
      * @returns {string}
      */
     run(input, args) {
+        input = new Uint8Array(input);
         const [
                 keyLength,
                 sampleLength,