+++ tests/operations/index.mjs
@@ -86,6 +86,8 @@ import "./tests/Typex";
 import "./tests/BLAKE2b";
 import "./tests/BLAKE2s";
 import "./tests/Protobuf";
+import "./tests/ParseSSHHostKey";
+import "./tests/DefangIP";
 
 // Cannot test operations that use the File type yet
 //import "./tests/SplitColourChannels";